Refactor to use single fsspec config for connectors
