+++ unstructured/ingest/runner/box.py
@@ -17,7 +17,7 @@ class BoxRunner(FsspecBaseRunner):
         self.read_config.download_dir = update_download_dir_remote_url(
             connector_name="box",
             read_config=self.read_config,
-            remote_url=self.fsspec_config.remote_url,
+            remote_url=self.fsspec_config.remote_url,  # type: ignore
             logger=logger,
         )
 
@@ -26,8 +26,8 @@ class BoxRunner(FsspecBaseRunner):
         source_doc_connector = BoxSourceConnector(  # type: ignore
             read_config=self.read_config,
             connector_config=SimpleBoxConfig(
-                path=self.fsspec_config.remote_url,
-                recursive=self.fsspec_config.recursive,
+                remote_url=self.fsspec_config.remote_url,  # type: ignore
+                recursive=self.fsspec_config.recursive,  # type: ignore
                 access_kwargs={"box_app_config": box_app_config},
             ),
             processor_config=self.processor_config,
