+++ unstructured/ingest/runner/gcs.py
@@ -17,7 +17,7 @@ class GCSRunner(FsspecBaseRunner):
         self.read_config.download_dir = update_download_dir_remote_url(
             connector_name="gcs",
             read_config=self.read_config,
-            remote_url=self.fsspec_config.remote_url,
+            remote_url=self.fsspec_config.remote_url,  # type: ignore
             logger=logger,
         )
 
@@ -25,8 +25,8 @@ class GCSRunner(FsspecBaseRunner):
 
         source_doc_connector = GcsSourceConnector(  # type: ignore
             connector_config=SimpleGcsConfig(
-                path=self.fsspec_config.remote_url,
-                recursive=self.fsspec_config.recursive,
+                remote_url=self.fsspec_config.remote_url,  # type: ignore
+                recursive=self.fsspec_config.recursive,  # type: ignore
                 access_kwargs={"token": token},
             ),
             read_config=self.read_config,
